Move methods from PullToRefreshListView to PullToRefreshListViewBase This makes it easier to extend PullToRefresh with a custom ListView
